Adjust module and imports
