+++ crates/typst-ide/src/tooltip.rs
@@ -3,8 +3,7 @@ use std::fmt::Write;
 use ecow::{eco_format, EcoString};
 use if_chain::if_chain;
 use typst::doc::Frame;
-use typst::eval::repr::Repr;
-use typst::eval::{CapturesVisitor, CastInfo, Tracer, Value};
+use typst::eval::{CapturesVisitor, CastInfo, Repr, Tracer, Value};
 use typst::geom::{round_2, Length, Numeric};
 use typst::syntax::ast::{self, AstNode};
 use typst::syntax::{LinkedNode, Source, SyntaxKind};
