+++ crates/typst-library/src/prelude.rs
@@ -16,7 +16,7 @@ pub use typst::doc::*;
 #[doc(no_inline)]
 pub use typst::eval::{
     array, cast, dict, format_str, func, scope, ty, Args, Array, Bytes, Cast, Dict,
-    FromValue, Func, IntoValue, Scope, Str, Symbol, Type, Value, Vm,
+    FromValue, Func, IntoValue, Repr, Scope, Str, Symbol, Type, Value, Vm,
 };
 #[doc(no_inline)]
 pub use typst::geom::*;
