+++ crates/typst/src/eval/none.rs
@@ -3,9 +3,8 @@ use std::fmt::Debug;
 
 use serde::{Serialize, Serializer};
 
-use super::{cast, ty, CastInfo, FromValue, IntoValue, Reflect, Type, Value};
+use super::{cast, ty, CastInfo, FromValue, IntoValue, Reflect, Repr, Type, Value};
 use crate::diag::StrResult;
-use crate::eval::repr::Repr;
 
 /// A value that indicates the absence of any other value.
 ///
