+++ crates/typst/src/eval/ops.rs
@@ -4,9 +4,8 @@ use std::cmp::Ordering;
 
 use ecow::eco_format;
 
-use super::{format_str, IntoValue, Regex, Value};
+use super::{format_str, IntoValue, Regex, Repr, Value};
 use crate::diag::{bail, StrResult};
-use crate::eval::repr::Repr;
 use crate::geom::{Align, Length, Numeric, Rel, Smart, Stroke};
 use Value::*;
