+++ crates/typst/src/model/introspect.rs
@@ -11,8 +11,7 @@ use indexmap::IndexMap;
 use super::{Content, Selector};
 use crate::diag::{bail, StrResult};
 use crate::doc::{Frame, FrameItem, Meta, Position};
-use crate::eval::repr::Repr;
-use crate::eval::{cast, func, scope, ty, Dict, Value, Vm};
+use crate::eval::{cast, func, scope, ty, Dict, Repr, Value, Vm};
 use crate::geom::{Point, Transform};
 use crate::model::Label;
 use crate::util::NonZeroExt;
